layout/base/tests/test_bug518777.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=518777
-->
<head>
  <title>Test for Bug 518777</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script>
  SimpleTest.waitForExplicitFinish();

  function dotest() {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

    var source = document.getElementById("source").contentWindow;
    var canvas = document.getElementById("scratch");
    var context = canvas.getContext("2d");

    context.clearRect(0, 0, canvas.width, canvas.height);
    context.drawWindow(source, 25, 25, canvas.width, canvas.height,
                       "transparent");

    var data = context.getImageData(0, 0, canvas.width, canvas.height).data;
    var i;
    for (i = 0; i < data.length; i += 4) {
      if (data[i] != 0 || data[i + 1] != 0 || data[i + 2] != 255 || data[i + 3] != 255)
        break;
    }
    ok(i >= data.length, "all pixels blue");

    SimpleTest.finish();
  }
  </script>
</head>
<body>
<p><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=518777">Mozilla Bug 518777</a></p>

<!--
iframe source is
<html><body onload='window.scrollTo(0,99999999); document.documentElement.offsetWidth; window.parent.dotest();' style='background: rgb(0,0,255); width: 100px; height: 50100px;'></body></html>
-->
<iframe id="source" width="50" height="50"
       src="data:text/html,%3Chtml%3E%3Cbody%20onload%3D%27window.scrollTo%280%2C99999999%29%3B%20document.documentElement.offsetWidth%3B%20window.parent.dotest%28%29%3B%27%20style%3D%27background%3A%20rgb%280%2C0%2C255%29%3B%20width%3A%20100px%3B%20height%3A%2050100px%3B%27%3E%3C%2Fbody%3E%3C%2Fhtml%3E"></iframe>
<canvas id="scratch" width="50" height="50"></canvas>

<pre id="test">
</pre>
</body>
</html>
